+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 - `pcs resource update` and `pcs resource meta` commands change master
   resources to promotable clone resources because master resources are
   deprecated in Pacemaker 2.x ([rhbz#1542288])
+- Support for the `promoted-max` bundle option replacing the `masters` option
+  in Pacemaker 2.x ([rhbz#1542288])
 
 ### Fixed
 - `pcs cluster cib-push diff-against=` does not consider an empty diff as
@@ -123,6 +125,10 @@
   ([rhbz#1557366])
 - CVE-2018-1079: Privilege escalation via authorized user malicious REST call
   ([rhbz#1550243])
+
+### Deprecated
+- The `masters` bundle option is obsoleted by the `promoted-max` option
+  in Pacemaker 2.x and therefore in pcs ([rhbz#1542288])
 
 [ghpull#166]: https://github.com/ClusterLabs/pcs/pull/166
 [rhbz#1158816]: https://bugzilla.redhat.com/show_bug.cgi?id=1158816